Relax libldb BuildRequires
Originally, we tried to stay on the safe side with libldb since it never
really commited to stable ABI or API, but since there were never any
issues in many years, it's safe to relax the requirement.

This change will benefit especially the storage developers who often
need a different (typically newer) libldb version and would like to
avoid to rebuild sssd for no reason.